refactor(pricing): all monetary amounts are integer yuan
Round all money fields (unit_price, subtotal, cost_unit_price,
cost_subtotal, profit, totals) to int via ROUND_HALF_UP instead
of keeping two decimal places. Percentages (margin_pct,
profit_rate) and deal-price factors remain float.

- pricing.round_money() now returns int
- QuoteItemBreakdown money fields typed as int
- render_pdf.fmt_money / Decimal quantize use integer precision
- render_xlsx number_format #,##0 (no decimals), float->int

+++ app/domain/schema.py
@@ -99,11 +99,11 @@ class QuoteItemBreakdown(BaseModel):
     unit: str
     qty: int
     list_price: int
-    unit_price: float
-    subtotal: float
-    cost_unit_price: float
-    cost_subtotal: float
-    profit: float
+    unit_price: int
+    subtotal: int
+    cost_unit_price: int
+    cost_subtotal: int
+    profit: int
     margin_pct: float
     protected: bool
     factor: float
